cdk/requirements.txt
aws-cdk-lib
constructs
boto3 